# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Pusk API
  version: 0.7.0
  description: |
    Self-hosted alert platform for ops teams. Telegram Bot API compatible.

    ## Authentication

    Three auth schemes are used depending on the endpoint group:

    - **Bot Token** — passed as a path parameter in `/bot/{token}/{method}` endpoints.
    - **JWT Bearer** — returned by `/api/auth` or `/api/register`, used for all `/api/` client endpoints.
    - **Admin Token** — server-configured token for `/admin/` management endpoints. Some admin endpoints also accept JWT with admin role.
  license:
    name: BSL 1.1
    url: https://github.com/getpusk/pusk/blob/main/LICENSE
  contact:
    name: Pusk
    url: https://getpusk.ru
servers:
  - url: https://getpusk.ru
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/auth or /api/register. 7-day TTL.
    AdminToken:
      type: apiKey
      in: header
      name: Authorization
      description: Admin token configured on the server. Passed as `Bearer {admin_token}`.
  schemas:
...
    Channel:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        bot_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    Bot:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        username:
          type: string
        token:
          type: string
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
...
paths:
...
  /admin/bots:
    post:
      tags: [Admin]
      operationId: adminCreateBot
      summary: Register a new bot
      security:
        - AdminToken: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, name]
              properties:
                token:
                  type: string
                  description: Bot API token (generated or provided).
                name:
                  type: string
      responses:
        '200':
          description: Bot created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bot'
  /admin/bots/{botID}:
    put:
      tags: [Admin]
      operationId: adminRenameBot
      summary: Rename a bot
      security:
        - AdminToken: []
        - BearerAuth: []
      parameters:
        - name: botID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Bot renamed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /admin/channel:
    post:
      tags: [Admin]
      operationId: adminCreateChannel
      summary: Create a channel via admin API
      security:
        - AdminToken: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Channel created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                  result:
                    $ref: '#/components/schemas/Channel'
  /admin/channel/{channelID}:
    put:
      tags: [Admin]
      operationId: adminRenameChannel
      summary: Rename a channel
      security:
        - AdminToken: []
        - BearerAuth: []
      parameters:
        - name: channelID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Channel renamed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
    delete:
      tags: [Admin]
      operationId: adminDeleteChannel
      summary: Delete a channel
      security:
        - AdminToken: []
        - BearerAuth: []
      parameters:
        - name: channelID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Channel deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /admin/reset-password:
    post:
      tags: [Admin]
      operationId: adminResetPassword
      summary: Reset a user's PIN (admin token only)
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org, username, new_pin]
              properties:
                org:
                  type: string
                username:
                  type: string
                new_pin:
                  type: string
      responses:
        '200':
          description: Password reset.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /admin/set-role:
    post:
      tags: [Admin]
      operationId: adminSetRole
      summary: Set a user's role (admin token only)
      security:
        - AdminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [org, user_id, role]
              properties:
                org:
                  type: string
                user_id:
                  type: integer
                  format: int64
                role:
                  type: string
                  enum: [admin, member]
      responses:
        '200':
          description: Role updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /api/org/register:
    post:
      tags: [Admin]
      operationId: adminOrgRegister
      summary: Register a new organization
      description: Creates a new org with the first user as admin. No auth required (self-service org creation).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, username, pin]
              properties:
                slug:
                  type: string
                  description: URL-safe org identifier.
                name:
                  type: string
                  description: Human-readable org name.
                username:
                  type: string
                pin:
                  type: string
      responses:
        '200':
          description: Org created, first user authenticated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  org:
                    type: string
                  token:
                    type: string
                  user_id:
                    type: integer
                    format: int64
                  username:
                    type: string
                  role:
                    type: string
                    enum: [admin]
        '409':
          description: Org slug already exists.